chore: achieve zero-warning build across all project modules
This comprehensive update resolves over 100 compiler warnings across core,
UI, and syntax components, ensuring a clean build and improved code quality.

Key changes include:
- Resolved unused parameter warnings using G_GNUC_UNUSED across all
  signal handlers and callbacks (affects 19 files).
- Fixed ISO C99 overlength string limit issues in main.c by splitting
  the Shortcuts UI XML into multiple literals and concatenating at runtime.
- Addressed 'ISO C forbids conversion of object pointer to function pointer'
  warnings in editor-widget.c using union-based casting compatible with -Wpedantic.
- Corrected missing initializers for GActionEntry structures in main.c,
  editor-input.c, and status-bar.c.
- Fixed signed/unsigned comparison warnings in editor-scrolling.c,
  editor-actions.c, and editor-input.c.
- Cleaned up dead code and unused variables in document.c, piece-table.c,
  and syntax files (python, shell, yaml, desktop).
- Resolved undeclared variable usage in document.c and editor-input.c
  during refactoring.

Verified with a full clean build: ninja -C builddir clean && ninja -C builddir.

+++ src/editor-actions.c
@@ -476,7 +476,7 @@ editor_widget_indent_selection(EditorWidget *self)
     
     document_begin_undo_group(self->doc);
     
-    for (int i = lines->len - 1; i >= 0; i--) {
+    for (int i = (int)lines->len - 1; i >= 0; i--) {
         size_t line_idx = g_array_index(lines, size_t, i);
         size_t line_start = document_get_offset_of_line(self->doc, line_idx);
         
@@ -540,7 +540,7 @@ editor_widget_unindent_selection(EditorWidget *self)
     
     document_begin_undo_group(self->doc);
     
-    for (int i = lines->len - 1; i >= 0; i--) {
+    for (int i = (int)lines->len - 1; i >= 0; i--) {
         size_t line_idx = g_array_index(lines, size_t, i);
         size_t line_off = document_get_offset_of_line(self->doc, line_idx);
         
@@ -556,7 +556,7 @@ editor_widget_unindent_selection(EditorWidget *self)
                 delete_len = 1;
             } else if (line_text[0] == ' ') {
                 size_t spaces = 0;
-                while (spaces < self->indent_width && spaces < len && line_text[spaces] == ' ') {
+                while (spaces < (size_t)self->indent_width && spaces < len && line_text[spaces] == ' ') {
                     spaces++;
                 }
                 if (spaces > 0) {
